Replace some backreferences in regexps, for portability.
The first is replaced with an identical expression; the other two could be done the same way, if desired, but `/['"]use strict['"]/` seems acceptable to me, even though technically it matches `'use strict"` and `"use strict'` which we don't really want. They are, however, so unlikely to occur in valid Erlang *or* ECMAScript code that I don't think it matters.

`\1` appears in two other rules that are too complicated for me to change.