Change some regexps using lookahead to avoid it, for portability.
The behavior of the regexps is unchanged. In the case of the BrighterScript regexps, it looks like the original regexps may not do what they were meant to do, but I decided not to try to fix them in this PR.

Lookahead is used in one remaining rule `'.txt' -> Vim Help File` that is too complicated for me to change.